Correct naming
Correct naming
remove old class chirk list
correct structur
