+++ lib/entity/user.dart
