+++ lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:chirk/pages/homePage.dart';
-import 'package:chirk/pages/loginPage.dart';
+import 'package:chirk/widget/home_widget.dart';
+import 'package:chirk/widget/login_widget.dart';
 import 'package:chirk/theme/theme.dart';
 import 'package:easy_dynamic_theme/easy_dynamic_theme.dart';
 import 'package:flutter/material.dart';
@@ -29,8 +29,8 @@ class _ChirkAppState extends State<ChirkApp> {
       themeMode: EasyDynamicTheme.of(context).themeMode,
       initialRoute: '/',
       routes: {
-        '/': (context) => const HomePage(),
-        '/login': (context) => LoginPage(),
+        '/': (context) => const HomeWidget(),
+        '/login': (context) => LoginWidget(),
       },
     );
   }
