+++ lib/service/chirk_list_service_list/ichirk_list_service_list.dart
@@ -1,0 +1,12 @@
+import 'dart:math';
+
+import 'package:flutter/cupertino.dart';
+
+import '../../entity/chirk.dart';
+import '../../entity/user.dart';
+import '../chirl_list_service.dart';
+
+abstract class IChirkListServiceList implements IChirkListService {
+  @protected
+  List<Chirk> initChirkList();
+}
