+++ lib/service/chirl_list_service.dart
@@ -1,0 +1,11 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+
+import '../entity/chirk.dart';
+import '../entity/user.dart';
+
+abstract class IChirkListService{
+  List<Chirk> getChirks();
+}
+
