+++ lib/widget/profileWidget.dart
@@ -1,8 +1,9 @@
 import 'package:easy_dynamic_theme/easy_dynamic_theme.dart';
 import 'package:flutter/material.dart';
 
-import '../../models/user.dart';
-import '../userIcons.dart';
+import '../entity/user.dart';
+import '../service/userIcons.dart';
+
 
 class ProfileWidget extends StatefulWidget {
   final User _user;
